refactoring stock transfer to follow controller pattern
